Selection box transforms into a bracket
When dragging mouse to create selection box and one of the axis is less than minimum length than the box changes to a double ended line (like a bracket).
MinSelectionBox size is added to InputMap with a default value of 5 (instead of 2)
also add the ability to change MinSelectionBox iin the demo window.